# Unify representation of tuples and other generic types