--- a/tests/baselines/reference/iterableArrayPattern29.errors.txt
+++ b/tests/baselines/reference/iterableArrayPattern29.errors.txt
@@ -1,7 +1,6 @@
 tests/cases/conformance/es6/destructuring/iterableArrayPattern29.ts(1,33): error TS2501: A rest element cannot contain a binding pattern.
 tests/cases/conformance/es6/destructuring/iterableArrayPattern29.ts(2,21): error TS2345: Argument of type '[string, boolean]' is not assignable to parameter of type '[string, number]'.
-  Types of property '1' are incompatible.
-    Type 'boolean' is not assignable to type 'number'.
+  Type 'boolean' is not assignable to type 'number'.
 
 
 ==== tests/cases/conformance/es6/destructuring/iterableArrayPattern29.ts (2 errors) ====
@@ -11,5 +10,4 @@ tests/cases/conformance/es6/destructuring/iterableArrayPattern29.ts(2,21): error
     takeFirstTwoEntries(...new Map([["", true], ["hello", true]]));
                         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 !!! error TS2345: Argument of type '[string, boolean]' is not assignable to parameter of type '[string, number]'.
-!!! error TS2345:   Types of property '1' are incompatible.
-!!! error TS2345:     Type 'boolean' is not assignable to type 'number'.
+!!! error TS2345:   Type 'boolean' is not assignable to type 'number'.