… tests/baselines/reference/destructuringParameterProperties5.errors.txt
… tests/baselines/reference/destructuringVariableDeclaration2.errors.txt
… tests/baselines/reference/genericCallWithTupleType.errors.txt
… tests/baselines/reference/optionalBindingParameters1.errors.txt
… tests/baselines/reference/optionalBindingParametersInOverloads1.errors.txt
… tests/baselines/reference/tupleTypes.errors.txt