With this PR we unify the internal representation of tuple types and other generic types. Tuple types are now represented as regular type references to synthesized generic types of the form:

```ts
interface Tuple<T0, T1, T2, ...> extends Array<T0 | T1 | T2 | ...> { 0: T0, 1: T1, 2: T2, ... }
```

With these changes it is no longer necessary to special case tuple types throughout the type checker. All of the optimizations we have in place for regular generic type references now also apply to tuple types.

Fixes #10431. The compile time for the example drops from 1.2s to 0.08s with `--lib es5` and from 30.5s to 0.11s with `--lib es6`.