@HerringtonDarkholme 2016-08-22
Will this be developed into https://github.com/Microsoft/TypeScript/issues/5453 ?

 Synthesized generic types and tuple types looks much like variadic kinds.