@weswigham 2016-08-22
I quite like the changes in the error baselines. Feels neater.